--- pkg/installer/oneagent/sudo_unix.go
@@ -1,9 +1,0 @@
-//go:build !windows
-
-package oneagent
-
-import "os"
-
-func needsSudo() bool {
-	return os.Getuid() != 0
-}
