--- pkg/installer/oneagent/sudo_windows.go
@@ -1,9 +1,0 @@
-//go:build windows
-
-package oneagent
-
-// needsSudo always returns false on Windows — the installer exe handles
-// privilege elevation itself via its embedded manifest.
-func needsSudo() bool {
-	return false
-}
