+++ pkg/installer/sudo_unix.go
@@ -4,8 +4,8 @@ package installer
 
 import "os"
 
-// needsSudo returns true when the current process is not running as root,
+// NeedsSudo returns true when the current process is not running as root,
 // indicating that the installer needs to be prefixed with sudo.
-func needsSudo() bool {
+func NeedsSudo() bool {
 	return os.Getuid() != 0
 }
