+++ pkg/installer/sudo_windows.go
@@ -2,8 +2,8 @@
 
 package installer
 
-// needsSudo always returns false on Windows — the installer exe handles
+// NeedsSudo always returns false on Windows — the installer exe handles
 // privilege elevation itself via its embedded manifest.
-func needsSudo() bool {
+func NeedsSudo() bool {
 	return false
 }
